Cleanup and rearranged function in utils
